studied if, elif, and else statement
You can use elif statement as many times as you want.

The order are important in if statements, the first order are considered highest priority. 
(Example)
If age >= 18:
   print("You are now signed up!")
elif age >= 100:
   print("You are too old to sign up")

The result of this code will still run "You are now signed up!" when you typed in 101 as your age, this is because the upper code will take priority if they are both met in conditions.


Note: Make sure that the lines under if statements are indented, this will indicated that what code should if statement run if they met the conditions.