Studied logic operators (or, and, not)
Logical operator (or) will run the code if one of the conditions were met
Ex:  eggs = 20
      is_cooked = False
      if egg == 20 or is_cooked == True:
      print("There might be some eggs in the fridge)
The result will print "There might be some eggs in the fridge" if any of the conditions were met.

Logical operator (and) will run the code if all of the and conditions were met
Ex:  boiled_beans = 4
      frequent_eat = "mung"
      if boiled beans  == 4 and frequent_eat == "red":
      print("I love eating red beans")
      else:
      print("I love eating mung beans")
The result of this code will print "I love eating mung beans, since all and conditions were not met.

Logical operator (not) will run the code if the not conditions were not false or not true
Ex:
     apple = "tasty"
     high_cal = False
     if apple == "tasty" and not high_cal == True:
     print("Apple is not only tasty, but also low in cal")
Note: Logical operator (not) can only be used with or/and logical operator. Ex: or not/ and not
The result of this code will print "Apple is not only tasty, but also low in cal" because the (not high_cal == True) means high_cal == False and thus condition were met.